requirements.txt
bs4
lxml
requests-toolbelt